[fix-add-reviews] * Fix delete sql requests * Fix filed types on Review and ReviewDO * Remove redundant access modifiers on FilmService interface * Fix logic to get all reviews

+++ src/main/java/ru/yandex/practicum/filmorate/domain/ReviewDO.java
@@ -8,8 +8,8 @@ import lombok.Value;
 @Builder(toBuilder = true)
 public class ReviewDO {
     @NonNull Long reviewId;
-    @NonNull Integer userId;
-    @NonNull Integer filmId;
+    @NonNull Long userId;
+    @NonNull Long filmId;
     String content;
     boolean isPositive;
     int useful;
